Move webhook notification test to separate module
